backend/pyproject.toml
[project]
name = "docflow-backend"
version = "0.1.0"
description = "DocFlow KZ - Document automation platform"
requires-python = ">=3.12"

dependencies = [
    # ─── Web Framework ───────────────────────────────────────
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # ─── Database ────────────────────────────────────────────
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "greenlet>=3.0.0",

    # ─── Validation ──────────────────────────────────────────
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "email-validator>=2.2.0",

    # ─── Auth ────────────────────────────────────────────────
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",

    # ─── Celery / Redis ──────────────────────────────────────
    "celery>=5.4.0",
    "redis>=5.0.0",
    "flower>=2.0.0",

    # ─── File Storage (MinIO/S3) ─────────────────────────────
    "boto3>=1.35.0",
    "minio>=7.2.0",

    # ─── Document Generation ─────────────────────────────────
    "weasyprint>=62.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "jinja2>=3.1.0",

    # ─── Excel/1C Import ─────────────────────────────────────
    "pandas>=2.2.0",
    "xlrd>=2.0.0",

    # ─── Email ───────────────────────────────────────────────
    "fastapi-mail>=1.4.0",

    # ─── HTTP Client ─────────────────────────────────────────
    "httpx>=0.27.0",

    # ─── Utils ───────────────────────────────────────────────
    "python-slugify>=8.0.0",
    "Pillow>=10.4.0",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
